requirements.txt
fastapi~=0.110.2
pydantic~=2.7.1
uvicorn~=0.29.0
aiohttp~=3.9.5
python-dotenv~=1.0.1
SQLAlchemy~=2.0.29
asyncpg~=0.29.0